Apply a patch for 64 bit Windows build. See JIRA issue XERCESC-1531 for details. Thanks to Boris Kolpackov
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/branches/xerces-2.7@568012 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/internal/XMLScanner.hpp
@@ -1298,7 +1298,7 @@ inline void XMLScanner::setValidateAnnotations(const bool newValue)
 inline void XMLScanner::setInputBufferSize(const size_t bufferSize)
 {
     fBufferSize = bufferSize;
-    fCDataBuf.setFullHandler(this, fBufferSize);
+    fCDataBuf.setFullHandler(this, (unsigned int)fBufferSize);
 }
 
 inline void XMLScanner::setIgnoredCachedDTD(const bool newValue)
